refactor : Line 리팩토링 Line의 생성을 Participant로 이동 LineTest의 크기 테스트 코드 변경
