+++ src/main/java/nextstep/ladder/domain/Point.java
@@ -1,0 +1,8 @@
+package nextstep.ladder.domain;
+
+import java.util.List;
+
+public interface Point {
+
+    List<Boolean> generatePoints();
+}
