--- src/main/java/nextstep/ladder/domain/PointGenerator.java
@@ -1,6 +1,0 @@
-package nextstep.ladder.domain;
-
-public interface PointGenerator {
-
-    Boolean generate();
-}
